requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
